Separate dev and prod settings in webpack
Replace settings.ts with a section in webpack.config.js that changes depending on which mode webpack is run in, and update `npm run dev` and `npm run build` to run in the appropriate modes. The `npm run watch` command might needs its mode set, but I don't know what that command is being used for, or even if it's still being used.

There are probably other settings in webpack.config.js that could be for development mode only (e.g. devServer), but I didn't mess with those for now.